+++ Cargo.toml
@@ -4,12 +4,12 @@ version = "0.1.0"
 authors = ["Raphaël Lustin <raphael@lustin.fr>"]
 
 [dependencies]
-clap = { version = "2.33.0", features = ["yaml"] }
-quick-xml = { version = "0.17", features = ["serialize"] }
-serde = { version = "1.0", features = ["derive"] }
-minimp3 = "0.3"
 audrey = "0.2"
+clap = { version = "2.33.0", features = ["yaml"] }
+minimp3 = "0.3"
+quick-xml = { version = "0.17", features = ["serialize"] }
 rayon = "1.3"
+serde = { version = "1.0", features = ["derive"] }
 
 [dev-dependencies]
 assert_cmd = "0.10"
